requirements.txt
streamlit
torch
torchvision
pandas
opencv-python-headless
scikit-learn
pydicom
transformers
sentencepiece
Pillow
PyYAML
transformers>=4.30.0
torch
streamlit
Pillow
numpy